Update vthunting.py
The limit length of message for Telegram API is 4096 characters. If it exceed this length, Telegram API will return error, thus no message will comes in. This code ensures that the messages you send are within the limits imposed by the Telegram API by splitting the result in "report" to 4096 characters.